restructure folders for better overview
