+++ src/display/DisplayPageManager.cpp
@@ -1,21 +1,10 @@
 #include "DisplayPageManager.h"
 
-// #include "../userConfig.h"
-// #if (DISPLAY_HARDWARE == 4)
-// #define TFT_eSPI_DISPLAY
-// #include "displayPageTypeEnumTFT.h"
-// #elif (DISPLAY_HARDWARE == 1) || (DISPLAY_HARDWARE == 2) || (DISPLAY_HARDWARE == 3)
-// #define U8G2_DISPLAY
-// #include "displayPageTypeOLEDEnum.h"
-// #else
-// #define NO_DISPLAY
-// #endif
-
-#include "DisplayPageStandard.h"
-#include "DisplayPageMinimal.h" 
-#include "DisplayPageTemperatureOnly.h" 
-// #include "DisplayPageScale.h" 
-#include "DisplayPageColor.h" 
+#include "templates/DisplayPageStandard.h"
+#include "templates/DisplayPageMinimal.h" 
+#include "templates/DisplayPageTemperatureOnly.h" 
+// #include "templates/DisplayPageScale.h" 
+#include "templates/DisplayPageColor.h" 
 
 DisplayPageManager::DisplayPageManager(DisplayManager *instanceOfDisplayManager) {
     displayManager = instanceOfDisplayManager;
