+++ src/display/templates/DisplayPageColor.cpp
