+++ src/display/templates/DisplayPageColor.h
