+++ src/display/templates/DisplayPageColorBase.cpp
