+++ src/display/templates/DisplayPageColorBase.h
