+++ src/display/templates/DisplayPageMinimal.cpp
