+++ src/display/templates/DisplayPageMinimal.h
