+++ src/display/templates/DisplayPageTemperatureOnly.cpp
