+++ src/display/templates/DisplayPageTemperatureOnly.h
