+++ src/display/templates/displayTemplateUpright.h
